Added confirmation boxes for deleting and stopping runs.  Extended the wait time between creating a run and showing the run details so that the run has time to start up before we try to display it

+++ pegasus/gtfar/static/js/controllers/runsController.js
@@ -89,8 +89,18 @@ function(moment) {
         };
 
         $scope.deleteSelected = function() {
+
             var selected = $scope.runsGrid.selectedItems;
             var deleteTotal = $scope.runsGrid.selectedItems.length;
+            var runName = selected[0].name;
+            if(deleteTotal > 1) {
+                runName = "these " + deleteTotal + " runs"
+            }
+            var doubleCheck = confirm("Are you sure you want to delete " + runName + "?  This will delete all output files as well.");
+            if(!doubleCheck){
+                return;
+            }
+
             var deleteCount = 0;
             for(var i = 0; i < deleteTotal; i++) {
                 $http.delete($window.apiLinks.runs + "/" + selected[i].id)
